Fixed casting pointer to int error. Fixed /**/ comment inside comment warnings. Fixed converting string constant to 'char*' warning.

+++ src/ExpToCasExe/expscan.cxx
@@ -15,6 +15,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <cstdint>
 
 #if defined(__clang__)
 #pragma GCC diagnostic ignored "-Wmissing-braces"
@@ -2771,7 +2772,7 @@ yylook()
 				}
 # endif
 			yyr = yyt;
-			if ( (int)yyt > (int)yycrank){
+			if ( (intptr_t)yyt > (intptr_t)yycrank){
 				yyt = yyr + yych;
 				if (yyt <= yytop && yyt->verify+yysvec == yystate){
 					if(yyt->advance+yysvec == YYLERR)	/* error transitions */
@@ -2785,7 +2786,7 @@ yylook()
 					}
 				}
 # ifdef YYOPTIM
-			else if((int)yyt < (int)yycrank) {		/* r < yycrank */
+			else if((intptr_t)yyt < (intptr_t)yycrank) {		/* r < yycrank */
 				yyt = yyr = yycrank+(yycrank-yyt);
 # ifdef LEXDEBUG
 				if(debug)fprintf(yyout,"compressed state\n");
